[MarathonSurvey] Move these as widgets
