+++ src/components/widgets/SurveyStatisticsList.vue
