+++ src/views/Home.vue
@@ -43,11 +43,11 @@ a {
 }
 </style>
 <script>
-import SurveyStatisticsList from '../components/marathon-survey/SurveyStatisticsList.vue'
-import LatestBreakPersonalBestList from '../components/marathon-survey/LatestBreakPersonalBestList.vue'
-import LatestCompleteFirstMarathonList from '../components/marathon-survey/LatestCompleteFirstMarathonList.vue'
-import LatestRegisteredRunnersList from '../components/marathon-survey/LatestRegisteredRunnersList.vue'
-import RunnersHonorRoll from '../components/marathon-survey/RunnersHonorRoll.vue'
+import SurveyStatisticsList from '../components/widgets/SurveyStatisticsList.vue'
+import LatestBreakPersonalBestList from '../components/widgets/LatestBreakPersonalBestList.vue'
+import LatestCompleteFirstMarathonList from '../components/widgets/LatestCompleteFirstMarathonList.vue'
+import LatestRegisteredRunnersList from '../components/widgets/LatestRegisteredRunnersList.vue'
+import RunnersHonorRoll from '../components/widgets/RunnersHonorRoll.vue'
 
 export default {
   name: "Home",
